add docbool/man pages

+++ install_locale.sh
@@ -8,6 +8,19 @@ if [ -z $APPLANG ]; then
 	exit 1
 fi
 
-install -d $BASEPATH/usr/share/locale/$APPLANG/LC_MESSAGES
-install --mode=644 locale/$APPLANG/LC_MESSAGES/*.mo $BASEPATH/usr/share/locale/$APPLANG/LC_MESSAGES
+if [ -e locale/$APPLANG/LC_MESSAGES ]; then
+	install -d $BASEPATH/usr/share/locale/$APPLANG/LC_MESSAGES
+	install --mode=644 locale/$APPLANG/LC_MESSAGES/*.mo $BASEPATH/usr/share/locale/$APPLANG/LC_MESSAGES
+fi
+
+if [ -e docbook/$APPLANG ]; then
+	install -d $BASEPATH/usr/share/gnome/help/backintime/$APPLANG/figures
+	install --mode=644 docbook/$APPLANG/*.xml $BASEPATH/usr/share/gnome/help/backintime/$APPLANG
+	install --mode=644 docbook/$APPLANG/figures/*.png $BASEPATH/usr/share/gnome/help/backintime/$APPLANG/figures
+fi
+
+if [ -e man/$APPLANG ]; then
+	install -d $BASEPATH/usr/share/man/$APPLANG/man1
+	install --mode=644 man/C/*.gz $BASEPATH/usr/share/man/$APPLANG/man1
+fi
